+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Link to tests. Fixes UICHKOUT-61.
 * Allow selection of more than one scan id. Fixes UICHKOUT-49.
 * Rewire links from items to inventory. Fixes UICHKOUT-48.
+* Remove settings from ui-checkout. Fixes UICHKOUT-64.
 
 ## [1.1.2](https://github.com/folio-org/ui-checkout/tree/v1.1.2) (2017-09-02)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v1.1.1...v1.1.2)